render_requirements.txt
# Render.com specific requirements
# Core dependencies
python-telegram-bot==13.15
flask==3.0.0
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Proxy support (if needed)
python-socks==2.4.4

# Markdown processing
markdown2==2.4.10

# Production server
gunicorn==21.2.0

# Image processing (required for telegram bot)
Pillow==10.0.1

# Additional system dependencies for Render
# These are automatically available on Render
# - imghdr (built-in Python module)
# - sqlite3 (built-in Python module)
# - json (built-in Python module)
# - datetime (built-in Python module)
# - threading (built-in Python module)
# - time (built-in Python module)
# - os (built-in Python module)
# - sys (built-in Python module)
# - signal (built-in Python module)
# - re (built-in Python module)
# - logging (built-in Python module)
# - argparse (built-in Python module)
